+++ src/Accord.Bot/Helpers/Permissions/DiscordPermissionType.cs
@@ -1,0 +1,9 @@
+namespace Accord.Bot.Helpers
+{
+    public enum DiscordPermissionType
+    {
+        Allow,
+        Deny,
+        All
+    }
+}
