+++ src/Accord.Bot/Helpers/TimeToMentionType.cs
@@ -1,0 +1,13 @@
+namespace Accord.Bot.Helpers
+{
+    public enum TimeToMentionType
+    {
+        ShortTime = 't',
+        LongTime = 'T',
+        ShortDate = 'd',
+        LongDate = 'D',
+        ShortDateTime = 'f',
+        LongDateTime = 'F',
+        RelativeTime = 'R'
+    }
+}
