+++ src/Accord.Bot/Infrastructure/BotState.cs
@@ -1,17 +1,8 @@
-using System;
-using System.Threading;
-
 namespace Accord.Bot.Infrastructure
 {
     public class BotState
     {
-        private long _isCacheReady;
-        public bool IsCacheReady
-        {
-            get => Interlocked.Read(ref _isCacheReady) == 1;
-            set => Interlocked.Exchange(ref _isCacheReady, Convert.ToInt64(value));
-        }
-
+        public bool IsCacheReady { get; set; }
         public bool IsReady => IsCacheReady;
     }
 }
